+++ sphinx/application/network/features/feature-network-relay-call/src/main/java/chat/sphinx/feature_network_relay_call/NetworkRelayCallErrorDto.kt
@@ -1,0 +1,11 @@
+package chat.sphinx.feature_network_relay_call
+
+import chat.sphinx.concept_network_relay_call.RelayResponse
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class NetworkRelayCallErrorDto(
+    override val success: Boolean,
+    override val response: Any?,
+    override val error: String?
+): RelayResponse<Any>()